% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_barplot.R
\name{rxpv.group_segment_plot}
\alias{rxpv.group_segment_plot}
\title{Grouped Segment Plot}
\usage{
rxpv.group_segment_plot(
  my_data,
  class_colname,
  Gp1_colname,
  Gp2_colname,
  order_by,
  point_size = 3,
  point_gp1_col = "white",
  point_gp2_col = "white",
  seg_size = 1,
  seg_gp1_col = "black",
  seg_gp2_col = "orange",
  ylim_gp1 = NULL,
  ylim_gp2 = NULL
)
}
\arguments{
\item{my_data}{a dataframe that contains \code{class_name}, \code{Gp1_colname} and \code{Gp2_colname}.}

\item{class_colname}{class column name}

\item{Gp1_colname}{group1 column name}

\item{Gp2_colname}{group2 column name}

\item{order_by}{ordered by gp1 or gp2 column name}

\item{point_size}{point size default 3}

\item{point_gp1_col}{gp1 color of point; default "white"}

\item{point_gp2_col}{gp2 color of point; default "white"}

\item{seg_size}{segment size}

\item{seg_gp1_col}{gp1 color of segment; default "black"}

\item{seg_gp2_col}{gp2 color of segment; default "orange"}

\item{ylim_gp1}{ylim for gp1}

\item{ylim_gp2}{ylim for gp2}
}
\value{
a grouped segment plot
}
\description{
Make a grouped segmnet plot
}
\examples{

## load example data

file_loc <- system.file('extdata', 'example_data.tsv', package = 'rxpViz', mustWork = TRUE)
dfX <- read.table(file_loc, sep="\t", stringsAsFactors = FALSE, header = 1)

## plot the grouped data
rxpv.group_segment_plot(my_data=dfX, class_colname="gene_id",
                      Gp1_colname="WT",
                      Gp2_colname="KO",
                      order_by="WT",
                      point_size=2,
                      point_gp1_col="white",
                      point_gp2_col="white",
                      seg_size=5,
                      seg_gp1_col="black",
                      seg_gp2_col="orange"
                      )

}
